Group all protocol related types into `protocol/`
See: https://github.com/sourcemeta/studio/issues/20
