+++ webview/src/components/FormatTab.tsx
@@ -1,4 +1,4 @@
-import type { CommandResult, FileInfo } from '../../../shared/types.ts';
+import type { CommandResult, FileInfo } from '../../../protocol/types';
 import { vscode } from '../vscode-api';
 import { RawOutput } from './RawOutput';
 import { Info, CheckCircle, AlertCircle, FileQuestion } from 'lucide-react';
