+++ webview/src/components/LintTab.tsx
@@ -1,4 +1,4 @@
-import type { LintResult } from '../../../shared/types.ts';
+import type { LintResult } from '../../../protocol/types';
 import { vscode } from '../vscode-api';
 import { RawOutput } from './RawOutput';
 import { CheckCircle, AlertCircle, FileQuestion } from 'lucide-react';
